pyproject.toml
[tool.poetry]
name = "syllabus-search-app-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
langchain = "^0.2.5"
openai = "^1.35.3"
chromadb = "^0.5.3"
langchain-openai = "^0.1.9"
langchain-chroma = "^0.1.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
langchain-community = "^0.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
